Align mw-ui-button with spec
* Introduce mw-ui-button.mw-ui-icon-element
In Vector, and Minerva we have buttons which are icon-only.
This seems to be the easiest way to achieve button focus states and
icon touch areas as we prepare to migrate this code in future
to WVUI

The states for progressive and destructive quiet buttons was
not correct so that is also fixed in this patchset.

Bug: T283757
Change-Id: I865688631b154f31420bcb265ffed89293da15a4