resourceloader: update some mw.loader documentation
For mw.loader#fnv132, the method returns a five-character
hash, not seven characters, since September 2019.

For mw.loader#registry, modules can be in a state of
'executing', which is listed and explained in the
documentation for each state, but not included in the
format details when listing all of the possible states.

Also tweak the order of states in the format details to match the
order they are documented in below.

Follow-up: I9ad11772a33b3a44cb625275b1d7353e1393ee49
Follow-up: I35c8d2bc2da7135456c39d4b513661d2d11fb0f4
Change-Id: Id27a08ec9d205248093ba158c0ea4a8677c50182