resourceloader: inline resolveIndexedDependencies
Merged into the only caller, mw.loader.register.

Previously, we would, for the initial registration of modules
with dependencies represented with index numbers instead
of names:

* call resolveIndexedDependencies, which would loop through
modules updating the dependencies with names
* do another loop through the modules, calling registerOne()

Instead, we can combine these loops. Its fine to call
registerOne() before the rest of the modules have their
dependencies converted from indexes to module names.

To improve readability, I flipped the order of the branches in
mw.loader.register and added an early return. Yes, this results
in an output size slightly bigger than using an `else` block
(7 characters for `return;` vs 6 for `else{}`) but it makes things
a lot easier to read.

Change-Id: I6c71ed70b3b308d7b2f67924637e23b08467035f