resourceloader: remove IIFE for mw.loader.store.requestUpdate
Declare the hasPendingWrites variable and the flushWrites
helper function in the outer scope for mw.loader. Its okay that they
are no longer scoped only to the requestUpdate() method, because the
entire file is wrapped in an IIFE, and nothing else in the file will
use them.

Remove the re-usable allocation for the onTimeout callback, and create
it on-demand it instead. This will only be created once for every
batch of pending writes (not for each write separately), so it is not
expensive to do.

Change-Id: Ieaed52611d79bcad6bd90c95d180cd035f75d255