resourceloader: add local variable for mw.loader.store
In mediawiki.loader.js, use a local variable for references to
mw.loader.store, which helps to both reduce the output size
and simplify interactions. Instead of a separate block for setting
mw.loader.store =, we can include `store: store` in the main
declaration of mw.loader, and then populate the object separately.

In flushWrites(), we now no longer need to define a local
variable `key` for `store.key`, because repeating the name
`store.key` results in a smaller output.

Should be a no-op in terms of functionality.

Change-Id: I33d372bdb71afceae3746cde0e2d51b2fb9bdb59