Use normalized CIDR range for navigation links on Special:Contributions
For the database query the cidr range is converted to hex values and a
valid, but wrong cidr is corrected without notice.
On the special page itself the links to other special pages using the
wrong cidr range, which leads to wrong or non-functional links.

For Special:Contribs/127.0.0.5/24 create links to other special pages
for ip range 127.0.0.0/24

Bug: T293664
Change-Id: I01051bee44ef263189f126fdd299674664ae5b63