registration: Apply manifest v2 logic to hypothetical v3
If a new manifest version would be published, this could end in a
regression if the corresponding if statements are forgotten to be
updated. If there is no breaking change in some area, the code from
manifest_version 2 should still be used. Therefore it makes sense to set
these if statements for all versions greater than 2 as well, to prevent
that when a new version is released, the code of version 1 is used.

Change-Id: I0a9feda19cec5be07e6ad0f46fef8c3140f14eca