resourceloader: pass $VARS.storeDisabled instead of .storeEnabled
When boolean values are minified, they get replaced by `!0` and
`!1` for true and false respectively. Since the mw.loader code wants
to check if the store is *disabled*, it uses `!$VARS.storeEnabled`,
which gets converted to either `!!0` or `!!1`. Inverting the variable
passed to the JavaScript code saves an extra character from the
response size, and makes the code a bit clearer.

The next version of the Minify library removes the conversion of
true and false to `!0` and `!1`, meaning that `!$VARS.storeEnabled`
will either be `!true` or `!false` - after this patch it become `false`
or `true` instead, resulting in the same number of characters in the
first case but two fewer in the second.

This should have no impact on the functionality.

Change-Id: I0cc3de3d2d43fc5ead65b471bc246dcd5bc4ae1e