profiler: Add 'running' option to ProfilerXhprof
If set to true, the enable function is not called.

For a few years now, WMF uses PHP auto_prepend file to start profiling
early so as to include the custom entrypoint wrapper (multiversion)
and the wmf-config logic that MediaWiki includes from LocalSettings
before profiling can start. Details at T180183.

However, this causes an issue with tideways-xhprof because when
MediaWiki enables the profiler a second time, it messed up the
internal state, as reported at T247332.

Bug: T247332
Change-Id: I8808cad03418d84ab9f36b884a2feef6dcc4abad