tests: suppress API prefix uniqueness check for 'pi'
This is shared between ProofreadPage's proofreadinfo and PageImages.

As these two extensions are not installed on the same wikis (PRP is
Wikisource-only, PageImages is disabled on Wikisource), there is not
urgent need to change one or the other, as that will break API
stability.

So, simply disable the API prefix uniqueness check for now (following
the example of T196962), and migrate the PRP info API as a follow up,
giving API clients time to update.

Bug: T290585
Change-Id: I7c35f911f254dcb7d2fa21bbac67416209229dcc
(cherry picked from commit d6db8f4ad0fca0ed904788f158df5d3d3a0576b3)