resourceloader: Optimize ES6 module detection in registerOne()
Instead of setting `requiresES6` to false, checking if the version
ends with an '!' (which indicates an ES6 only module), and then
if it does end with '!' setting requiresES6, just set requiresES6
directly to the result of the check, and use that variable as the
condition to remove the extra '!' from the end. Doing this saves
a few extra characters from the output size.

Should have no effect on functionality.

Change-Id: I8ad17a26b67ce4db0fb5524b63e4cae163b277e6