resourceloader: Move mw.loader.getModuleNames to mediawiki.base
It's not needed as part of the startup code, just for some modules
(in core the only place its used is in the mediawiki.inspect module)
and thus does not need to be defined so early, it can go in the
mediawiki.base module that is an implicit dependency of all
regular modules, where it can be loaded concurrently and take
advantage of the ResourceLoader module caching system.

Also tweak the documentation, the returned array is known to be
an array of strings.

Should have no effect on the functionality.

Change-Id: I18f5fae876a8b90f66b9c555d393169a5525f1af