resourceloader: update documentation for mw.loader.implement
If the script given is a string, this is executed via the `domEval()`
function - the documentation still referred to execution via
`$.globalEval()`. This was changed in 2018, when the startup code
was refactored to no longer depend on jQuery, see T192623.

Follow-up: Icdd5d76546a6c265a8e941c4e9b28f73bf9dd028
Change-Id: Ic3853a30ec50aafad55ac551a4848fe9db9b35cf