Installer: Add $wgLocaltimezone to LocalSettings template
If $wgLocaltimezone isn't hard-coded, it is evaluated from the
system timezone. On some systems this isn't guaranteed to be
static, for example on Debian, GMT can get converted to UTC,
instead of Europe/London. This can throw off code that relies
on consistent time zones or signatures. Hard-code the auto-
detected timezone value in LocalSettings.php to avoid this.

Also switch up the order of two variables to keep localization-
related settings next to each other.

Bug: T21991
Bug: T290464
Change-Id: I315a55a3129eeb81711232f59467eb8a2c59a80e