Make TitleValue fields private
Class is extended only in JsonConfig extension, but
these fields are not accessed.

The @deprecated annotation was put in there to prevent
external access, since the fields were protected we can't
use DeprecationHelper to hard deprecate. The class is
not @stable to extend, so for all intents and purposes
protected is equivalent to private.

Change-Id: I30089cf1f1361bf8f8c544316e422581ac7aea64