SkinFactory: Allow skins to be registered as "skippable"
The existing method of hardcoding ApiOutput and Fallback is not
great, and there is a use case inside Vector as we split
Vector skin into two as well as inside ContentTranslation.

This adds to the existing wgSkipSkins configuration by allowing
skins to register themselves as skippable without a LocalSettings
change.

Bug: T291193
Change-Id: I9caa8deb5b58fa1ef1eb548db497ef095cbbd154