profiler: Fixup broken logic for the new 'running' option
Follows-up I8808cad03418 which added this option, but I neglected to
actually test it. When trying it out with a patched wmf-config on
a mwdebug server, the following issue surfaced:

> Fatal TypeError: Argument 1 passed to XhprofData::__construct() must
> be of type array, null given, called in ProfilerXhprof.php:90

This is because ProfilerXhprof::disable() actually has its own state
tracking which refuses to even try calling tideways_xhprof_disable()
unless its state says it was in control of enabling it.

Remove this check and reduce the class to just abstracting the calls
to the PHP extension.

Also, remove the useless null return which is incompatible with the
callers to this function as evidenced by above Fatal TypeError.
The PHP extensions already default to an empty array, which is now
documented.

Test Plan:
* Local wiki with php-tideways installed, on this pach applied.
  Note that MediaWiki-Docker now has php-tideways pre-installed!
* In LocalSettings, configure profiling as per
  <https://www.mediawiki.org/wiki/MediaWiki-Docker/Recipes/Profiling>
  $wgProfiler = [
    'class'  => ProfilerXhprof::class,
    'output' => 'text',
  ];
* At load.php?forceprofile=1, there are valid percentages (100% "main").
* Add `tideways_xhprof_enable();` and `$wgProfiler['running'] = true;`
  to LocalSettings.
  The percentages are still valid, and the profile is now longer
  with additional coverage of early code such as MediaWikiServices.
* Switching to master, there is a fatal error with these settings.

Bug: T247332
Change-Id: Iaad88a0c46d39aee1374e5b02a77251bfa4c6f09
(cherry picked from commit c77f9a8c6092c9e000e14ed3abc8a6e44324fc6b)