EditPage: Read from wpWatchlistExpiry form value on "show changes" too
To allow people to set a custom watchlist expiry and then preview the
page, we read from the POSTed 'wpWatchlistExpiry' value and set that
as the default.

It seems reasonable and what users expect to do the same for "show changes",
so check for $this->diff as well.

Bug: T291287
Change-Id: Ia3b27b61a5c65b1830045796a87f11a5659502e2
(cherry picked from commit c27a559a934be146f2c7771ea0d75732f9cfd026)