resourceloader: remove `return` from mw.Message constructor
When invoked as a constructor, the created object gets returned
anyway. When not invoked as a construct (eg with
`var demo = mw.Message( mw.messages, 'brackets', [ 'param' ] )`
the `this` value is the global `mw` object, which is unexpected.

Using mw.Message as a method call is not supported, and it
causes other problems (the .format, .map, .key, and .parameters
properties get set on the global `mw` object) but those are left
to be dealt with separately.

This follows-up r89076 (6a408d444c1c4) where it was added for no
apparent reason.

Change-Id: Ie25758f92df0a33e66c266e007a1b81f077a9790