Revert "resourceloader: add local variable for mw.loader"
This reverts commit 0faa22fc6d3294a99ee622b3efbf258d9f7a9590.

Reason for revert: Gzip is usually (and in this case) more efficient at
natively deduplicating repeated strings when transferring code from
server to client, than is possible through manual abbreviating through
a variable.

Fresnel confirmed this patch reduces JS transfer size by 10B. Not much,
but enough for the optimisation to not be needed.

Change-Id: I4ad22ca1220f33f018a6bc7e3da8389246cf5f9c