Unbreak "show changes" in live preview for page creations
Switch live preview to use ApiComparePages, thus making it possible to
view changes for a page being created.

Additionally, ApiComparePages has a different behaviour for section
diffs, which is the same as EditPage: the content of the
(to|from)text-{slot} parameters is inserted as the given section for the
provided (to|from)title parameter.

Also trim the new text, for consistency with EditPage diffs.

Based on Ic87efe9f0140ef06c6be728839beb1f06930d3cc by TheDJ.

Bug: T109166
Change-Id: I635d3a9d59c25804fb71e7b31659f3e4cbc6ab5f