Use UndeletePage from ApiUndelete
UndeletePage is responsible for checking permissions. As a user-facing
change, if the user cannot undelete pages, the module will error with
the "permissiondenied" code and the "badaccess-groups" message. This is
consistent with other API modules (e.g. ApiDelete), and replaces the
previous behaviour of failing with code "permdenied-undelete" and no
message explaining the failure (which seems to be a bug). For everything
else, backwards compatibility is preserved; in particular, we still die
with "apierror-cantundelete" if nothing was deleted (which UndeletePage
considers a successful operation).

Bug: T290021
Change-Id: I89f5e9e4a1e120e6ca6dd52987059baefbcb84d5