Introduce a BlockRestrictionStoreFactory
The BlockRestrictionStoreFactory is a new service which allows fetching
of a wiki-specific BlockRestrictionStore by fetching the correct
LoadBalancer via a LBFactory and then injecting it into the
BlockRestrictionStore.

This is being done as a part of the project of supporting cross-wiki
blocking. Currently if a cross-wiki block with restrictions is being
created, the block is being stored in the foreign database, but the
restrictions are being stored in the local one.

Blocks were made wikiaware in I8ae8133f7e232cc75aae6b72fcd7feaeb313cba7.

Bug: T291983
Change-Id: Ia342030c31710f40c95acc75d3cc76f1ad79d806